update the cache instead of sending a second request 🤔

+++ src/DeleteButton.jsx
@@ -16,11 +16,20 @@ export default function DeleteButton({ id }) {
   const [deleteTodo] = useMutation(DELETE_TODOS);
 
   const hadnleDeleteTodo = async () => {
-    const deleteData = await deleteTodo({
-      variables: { id },
-      refetchQueries: [{ query: GET_TODOS }],
-    });
-    console.log("DELETE_TODOS", deleteData);
+    const isConfirmed = window.confirm(
+      "Are you sure you want to delete this todo?"
+    );
+    if (isConfirmed) {
+      const deleteData = await deleteTodo({
+        variables: { id },
+        update: (cache) => {
+          const prevData = cache.readQuery({ query: GET_TODOS });
+          const newTodos = prevData.todos.filter((todo) => todo.id !== id);
+          cache.writeQuery({ query: GET_TODOS, data: { todos: newTodos } });
+        },
+      });
+      console.log("DELETE_TODOS", deleteData);
+    }
   };
 
   return (
